Implement JPQL-query for fetching of all votes for today. This avoids db requests within a loop.

+++ src/main/java/ru/restaurants/restvoting/service/UserService.java
@@ -19,5 +19,5 @@ public interface UserService {
     /**
      *   Get all votes for custom date in a map like restaurantId -> votesTodayCount
      * */
-    Map<Integer, Integer> getVotesForToday();
+    Map<Integer, Long> getVotesForToday();
 }
